+++ tests/unit/test_cli.py
@@ -17,7 +17,7 @@ from libera_utils.aws import processing_step_function_trigger as psfn
          )
     ]
 )
-def test_parse_base_cli_args(cli_args, parsed):
+def test_parse_cli_args(cli_args, parsed):
     """
     Test that cli args are parsed properly
     """
@@ -45,7 +45,7 @@ def test_parse_base_cli_args(cli_args, parsed):
              overwrite=False))
     ]
 )
-def test_parse_make_kernel_cli_args(cli_args, parsed):
+def test_make_kernel_parse_cli_args(cli_args, parsed):
     """
     Test that cli args are parsed properly
     """
@@ -57,26 +57,73 @@ def test_parse_make_kernel_cli_args(cli_args, parsed):
 @pytest.mark.parametrize(
     ("cli_args", "parsed"),
     [
-        (['ecr-upload', 'fake_image', 'fake_tag', 'fake_algo', '--ignore-docker-config'],
-         argparse.Namespace(
-             func=ecr_upload.ecr_upload_cli_func,
-             image_name='fake_image',
-             image_tag='fake_tag',
-             algorithm_name='fake_algo',
-             ecr_image_tags=None,
-             ignore_docker_config=True)
-         ),
-        (['step-function-trigger', 'fake_algorithm', '2000-01-01', '--wait_for_finish', '--verbose'],
+        (['step-function-trigger', 'l1b-cam', '2030-01-01'],
          argparse.Namespace(
              func=psfn.step_function_trigger,
-             algorithm_name='fake_algorithm',
-             applicable_day='2000-01-01',
-             wait_for_finish=True,
-             verbose=True)
+             algorithm_name='l1b-cam',
+             applicable_day='2030-01-01',
+             wait_time=5)
+         ),
+        (['step-function-trigger', 'l1b-cam', '2030-01-01', '--wait-time=5',],
+         argparse.Namespace(
+             func=psfn.step_function_trigger,
+             algorithm_name='l1b-cam',
+             applicable_day='2030-01-01',
+             wait_time=5)
          )
     ]
 )
-def test_parse_aws_tools_cli_args(cli_args, parsed):
+def test_step_function_trigger_parse_cli_args(cli_args, parsed):
+    """
+    Test that cli args are parsed properly
+    """
+    print(f"CLI ARGS \n{cli_args}\n")
+    print(f"Parsed args: {parsed} \n")
+    assert cli.parse_cli_args(cli_args) == parsed
+
+
+@pytest.mark.parametrize(
+    ("cli_args", "parsed"),
+    [
+        (['ecr-upload', 'l1b-rad', 'test-image'],
+         argparse.Namespace(
+             func=ecr_upload.ecr_upload_cli_handler,
+             algorithm_name='l1b-rad',
+             image_name='test-image',
+             image_tag='latest',
+             ecr_tags=None,
+             ignore_docker_config=False)
+         ),
+        (['ecr-upload', 'l1b-rad', 'test-image', '--ignore-docker-config', '--image-tag', 'tag1.2'],
+         argparse.Namespace(
+             func=ecr_upload.ecr_upload_cli_handler,
+             algorithm_name='l1b-rad',
+             image_name='test-image',
+             image_tag='tag1.2',
+             ecr_tags=None,
+             ignore_docker_config=True)
+         ),
+        (['ecr-upload', 'l1b-rad', 'test-image', '--image-tag=latest'],
+         argparse.Namespace(
+             func=ecr_upload.ecr_upload_cli_handler,
+             image_name='test-image',
+             algorithm_name='l1b-rad',
+             image_tag='latest',
+             ecr_tags=None,
+             ignore_docker_config=False)
+         ),
+        (['ecr-upload', 'l1b-rad', 'test-image', '--ecr-tags', 'latest', 'tag2'],
+         argparse.Namespace(
+             func=ecr_upload.ecr_upload_cli_handler,
+             algorithm_name='l1b-rad',
+             image_name='test-image',
+             image_tag='latest',
+             ecr_tags=['latest', 'tag2'],
+             ignore_docker_config=False)
+         )
+    ]
+)
+def test_ecr_upload_cli_args(cli_args, parsed):
     """
     Test that cli args are parsed properly
     """
@@ -94,7 +141,7 @@ def test_parse_aws_tools_cli_args(cli_args, parsed):
              file_path='some/file/path.nc',
              algorithm_name='l1b-cam',
              account_suffix='-stage')),
-        (['s3-utils', 'put', 'l1b-cam', 'some/file/path.nc', '--account_suffix=-test'],
+        (['s3-utils', 'put', 'l1b-cam', 'some/file/path.nc', '--account-suffix=-test'],
          argparse.Namespace(
              func=s3_utilities.s3_put_cli_handler,
              file_path='some/file/path.nc',
@@ -105,7 +152,7 @@ def test_parse_aws_tools_cli_args(cli_args, parsed):
              func=s3_utilities.s3_list_cli_handler,
              algorithm_name='l1b-cam',
              account_suffix='-stage')),
-        (['s3-utils', 'ls', 'l1b-cam', '--account_suffix=-test'],
+        (['s3-utils', 'ls', 'l1b-cam', '--account-suffix=-test'],
          argparse.Namespace(
              func=s3_utilities.s3_list_cli_handler,
              algorithm_name='l1b-cam',